fix:friend-request alert flow
